feat: agregar servicio de cliente en el API
